[refactor] 파일명 변경, users 라우터 정보 변경
